requirements.txt
alembic==1.14.1
bcrypt==4.2.1
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
msgspec==0.19.0
packaging==24.2
psycopg2==2.9.10
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.1
redis==5.2.1
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3